isolating K.inverse()

+++ main.cu
@@ -78,7 +78,7 @@ int main()
         }
     }
     */
-    Eigen::MatrixXd Gradient = GetGradient(A, B , 1, 1);
+    Eigen::MatrixXf Gradient = GetGradient(A, B , 1, 1);
     for(int row = 0; row != Gradient.rows(); row++)
     {
         for(int col = 0; col != Gradient.cols(); col++)
